gpuworker/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
torch>=2.0.0
openai-whisper>=20231117
pydub>=0.25.1
python-multipart>=0.0.6
numpy>=1.24.0
typing-extensions>=4.8.0
ffmpeg-python>=0.2.0